Simplify target in justfile
Prior to this change, the target variable is hardcoded to
`aarch64-apple-darwin`. When we want to run `just test`
on a Linux environment, we must pass in the target.

This change modifies the `justfile` to dynamically determine the target
architecture using `rustc -Vv`. We do not need to pass the target
manually in any environment unless necessary.